Revert "refactor hooks"
This reverts commit b071beb350ff202dcd420ab4bfe0806b15a12249.
