+++ src/components/Accounts.tsx
@@ -20,7 +20,7 @@ import React, { useEffect, useState } from 'react';
 import styled from 'styled-components';
 
 import { useSourceTarget } from '../contexts/SourceTargetContextProvider';
-import { useAccounts } from '../hooks';
+import useAccounts from '../hooks/useAccounts';
 import formatAccounts from '../util/formatAccounts';
 import getChainSS58 from '../util/getSS58';
 import { Account, SubHeader } from '.';
