+++ src/components/Remark.tsx
@@ -19,7 +19,8 @@ import React, { useState } from 'react';
 import styled from 'styled-components';
 
 import { useTransactionContext } from '../contexts/TransactionContext';
-import { useLoadingApi, useSendMessage } from '../hooks';
+import useLoadingApi from '../hooks/useLoadingApi';
+import useSendMessage from '../hooks/useSendMessage';
 import { TransactionTypes } from '../types/transactionTypes';
 interface Props {
   className?: string;
