--- src/hooks/index.ts
@@ -1,30 +1,0 @@
-// Copyright 2021 Parity Technologies (UK) Ltd.
-// This file is part of Parity Bridges UI.
-//
-// Parity Bridges UI is free software: you can redistribute it and/or modify
-// it under the terms of the GNU General Public License as published by
-// the Free Software Foundation, either version 3 of the License, or
-// (at your option) any later version.
-//
-// Parity Bridges UI is distributed in the hope that it will be useful,
-// but WITHOUT ANY WARRANTY; without even the implied warranty of
-// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-// GNU General Public License for more details.
-//
-// You should have received a copy of the GNU General Public License
-// along with Parity Bridges UI.  If not, see <http://www.gnu.org/licenses/>.
-export { default as useAccounts } from './useAccounts';
-export { default as useApiConnection } from './useApiConnection';
-export { default as useBalance } from './useBalance';
-export { default as useBlocksInfo } from './useBlocksInfo';
-export { default as useBridgedBlocks } from './useBridgedBlocks';
-export { default as useConnectedReceiver } from './useConnectedReceiver';
-export { default as useDashboard } from './useDashboard';
-export { default as useDashboardProfile } from './useDashboardProfile';
-export { default as useDerivedAccount } from './useDerivedAccount';
-export { default as useLaneId } from './useLaneId';
-export { default as useLoadingApi } from './useLoadingApi';
-export { default as useMessagesLane } from './useMessagesLane';
-export { default as useSendMessage } from './useSendMessage';
-export { default as useTransactionPreparation } from './useTransactionPreparation';
-export { default as useTransactionType } from './useTransactionType';
