+++ src/hooks/useSendMessage.ts
@@ -25,7 +25,8 @@ import { useAccountContext } from '../contexts/AccountContextProvider';
 import { useUpdateMessageContext } from '../contexts/MessageContext';
 import { useSourceTarget } from '../contexts/SourceTargetContextProvider';
 import { useTransactionContext, useUpdateTransactionContext } from '../contexts/TransactionContext';
-import { useLaneId, useTransactionPreparation } from '../hooks';
+import useLaneId from '../hooks/useLaneId';
+import useTransactionPreparation from '../hooks/useTransactionPreparation';
 import { TransactionStatusEnum, TransactionTypes } from '../types/transactionTypes';
 import getSubstrateDynamicNames from '../util/getSubstrateDynamicNames';
 import logger from '../util/logger';
